+++ Source/EventFlow/ReadStores/IReadModelContextFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace EventFlow.ReadStores
+{
+    public interface IReadModelContextFactory
+    {
+        IReadModelContext Create(string readModelId, bool isNew);
+    }
+}