+++ Source/EventFlow/ReadStores/InMemory/InMemoryReadStore.cs
@@ -89,21 +89,25 @@ namespace EventFlow.ReadStores.InMemory
         }
 
         public override async Task UpdateAsync(IReadOnlyCollection<ReadModelUpdate> readModelUpdates,
-            Func<IReadModelContext> readModelContextFactory,
+            IReadModelContextFactory readModelContextFactory,
             Func<IReadModelContext, IReadOnlyCollection<IDomainEvent>, ReadModelEnvelope<TReadModel>, CancellationToken,
-            Task<ReadModelEnvelope<TReadModel>>> updateReadModel,
+                Task<ReadModelEnvelope<TReadModel>>> updateReadModel,
             CancellationToken cancellationToken)
         {
             using (await _asyncLock.WaitAsync(cancellationToken).ConfigureAwait(false))
             {
                 foreach (var readModelUpdate in readModelUpdates)
                 {
-                    if (!_readModels.TryGetValue(readModelUpdate.ReadModelId, out var readModelEnvelope))
+                    var readModelId = readModelUpdate.ReadModelId;
+
+                    var isNew = !_readModels.TryGetValue(readModelId, out var readModelEnvelope);
+
+                    if (isNew)
                     {
-                        readModelEnvelope = ReadModelEnvelope<TReadModel>.Empty(readModelUpdate.ReadModelId);
+                        readModelEnvelope = ReadModelEnvelope<TReadModel>.Empty(readModelId);
                     }
 
-                    var readModelContext = readModelContextFactory();
+                    var readModelContext = readModelContextFactory.Create(readModelId, isNew);
 
                     readModelEnvelope = await updateReadModel(
                         readModelContext,
@@ -115,11 +119,11 @@ namespace EventFlow.ReadStores.InMemory
 
                     if (readModelContext.IsMarkedForDeletion)
                     {
-                        _readModels.Remove(readModelUpdate.ReadModelId);
+                        _readModels.Remove(readModelId);
                     }
                     else
                     {
-                        _readModels[readModelUpdate.ReadModelId] = readModelEnvelope;
+                        _readModels[readModelId] = readModelEnvelope;
                     }
                 }
             }