+++ Source/EventFlow/ReadStores/ReadModelStore.cs
@@ -53,7 +53,7 @@ namespace EventFlow.ReadStores
             CancellationToken cancellationToken);
 
         public abstract Task UpdateAsync(IReadOnlyCollection<ReadModelUpdate> readModelUpdates,
-            Func<IReadModelContext> readModelContextFactory,
+            IReadModelContextFactory readModelContextFactory,
             Func<IReadModelContext, IReadOnlyCollection<IDomainEvent>, ReadModelEnvelope<TReadModel>, CancellationToken,
                 Task<ReadModelEnvelope<TReadModel>>> updateReadModel,
             CancellationToken cancellationToken);